Response Class created
Response class created… Responsible for all requests sent to the
server. Takes the place of all of the Resource and Service files.

+++ src/com/LakeShore/Model/Product.java
@@ -51,7 +51,7 @@ public class Product {
     
     public int getQuantity(){	return quantity;	}
     
-    public void setPrice(int x)	{	this.price = x;	}
+    public void setPrice(double x)	{	this.price = x;	}
     
     public double getPrice(){	return price;	}
