+++ src/com/LakeShore/service/representation/ProductRequest.java
@@ -43,7 +43,7 @@ public class ProductRequest {
     
     public int getQuantity(){	return quantity;	}
     
-    public void setPrice(int x)	{	this.price = x;	}
+    public void setPrice(double x)	{	this.price = x;	}
     
     public double getPrice(){	return price;	}
